Split Kokkos_SYCL_Parallel_Reduce.hpp
