+++ core/src/decl/Kokkos_Declare_SYCL.hpp
@@ -25,7 +25,8 @@
 #include <SYCL/Kokkos_SYCL_MDRangePolicy.hpp>
 #include <SYCL/Kokkos_SYCL_ParallelFor_Range.hpp>
 #include <SYCL/Kokkos_SYCL_ParallelFor_MDRange.hpp>
-#include <SYCL/Kokkos_SYCL_Parallel_Reduce.hpp>
+#include <SYCL/Kokkos_SYCL_ParallelReduce_Range.hpp>
+#include <SYCL/Kokkos_SYCL_ParallelReduce_MDRange.hpp>
 #include <SYCL/Kokkos_SYCL_Parallel_Scan.hpp>
 #include <SYCL/Kokkos_SYCL_Parallel_Team.hpp>
 #include <SYCL/Kokkos_SYCL_UniqueToken.hpp>
